Institution CRUD
Implemented the basic CRUD for Institution entity.
Changed the database concept, unique business id was added to the entity, a new migration and a database update was also made.
I've written an InitDb.sql script as well, to fill up my database with basic data.
Modified the Institution dtos, rewritten the HttpPost to HttpPatch in my ExaminerController, it was wrongly implemented.
Basic Program.cs and MappinService.cs modifications were made.