Settings page and login modification
I implemented the Settings page, we get back the data about the user with the GetMyProfile function which reads the user Id from the JWT token and gives back the data. I also modificated the login process, I have a dedicated button for it, I'm not navigating from the home page.